pyproject.toml
[project]
name = "projetoPI3"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "eduardafh", email = "eduardafheidemann@gmail.com"},
]
dependencies = ["django>=5.1.6", "djangorestframework>=3.15.2"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
